refactor: StudyCard 컴포넌트를 공용 컴포넌트로 전환 #1
